requirements.txt
runpod
openai
python-doctr[torch]
python-jose[cryptography]
torch